mpt: adjust verifyMPTProof naming and restructure index file (#3730)
