+++ packages/mpt/test/proof.spec.ts
@@ -7,7 +7,7 @@ import {
   createMPTFromProof,
   createMerkleProof,
   updateMPTFromMerkleProof,
-  verifyMPTProof,
+  verifyMerkleProof,
 } from '../src/index.js'
 
 describe('simple merkle proofs generation and verification', () => {
@@ -19,28 +19,28 @@ describe('simple merkle proofs generation and verification', () => {
     await trie.put(utf8ToBytes('key3cc'), utf8ToBytes('aVal3'))
 
     let proof = await createMerkleProof(trie, utf8ToBytes('key2bb'))
-    let val = await verifyMPTProof(utf8ToBytes('key2bb'), proof)
+    let val = await verifyMerkleProof(utf8ToBytes('key2bb'), proof)
     assert.equal(bytesToUtf8(val!), 'aVal2')
 
     proof = await createMerkleProof(trie, utf8ToBytes('key1aa'))
-    val = await verifyMPTProof(utf8ToBytes('key1aa'), proof)
+    val = await verifyMerkleProof(utf8ToBytes('key1aa'), proof)
     assert.equal(bytesToUtf8(val!), '0123456789012345678901234567890123456789xx')
 
     proof = await createMerkleProof(trie, utf8ToBytes('key2bb'))
-    val = await verifyMPTProof(utf8ToBytes('key2'), proof)
+    val = await verifyMerkleProof(utf8ToBytes('key2'), proof)
     // In this case, the proof _happens_ to contain enough nodes to prove `key2` because
     // traversing into `key22` would touch all the same nodes as traversing into `key2`
     assert.equal(val, null, 'Expected value at a random key to be null')
 
     let myKey = utf8ToBytes('anyRandomKey')
     proof = await createMerkleProof(trie, myKey)
-    val = await verifyMPTProof(myKey, proof)
+    val = await verifyMerkleProof(myKey, proof)
     assert.equal(val, null, 'Expected value to be null')
 
     myKey = utf8ToBytes('anotherGarbageKey') // should generate a valid proof of null
     proof = await createMerkleProof(trie, myKey)
     proof.push(utf8ToBytes('123456')) // extra nodes are just ignored
-    val = await verifyMPTProof(myKey, proof)
+    val = await verifyMerkleProof(myKey, proof)
     assert.equal(val, null, 'Expected value to be null')
 
     await trie.put(utf8ToBytes('another'), utf8ToBytes('3498h4riuhgwe')) // cspell:disable-line
@@ -49,7 +49,7 @@ describe('simple merkle proofs generation and verification', () => {
     proof = await createMerkleProof(trie, utf8ToBytes('another'))
     // and try to use that proof on another key
     try {
-      await verifyMPTProof(utf8ToBytes('key1aa'), proof)
+      await verifyMerkleProof(utf8ToBytes('key1aa'), proof)
       assert.fail('expected error: Invalid proof provided')
     } catch (e: any) {
       assert.equal(e.message, 'Invalid proof provided')
@@ -59,7 +59,7 @@ describe('simple merkle proofs generation and verification', () => {
     proof = await createMerkleProof(trie, utf8ToBytes('key2bb'))
     proof[0].reverse()
     try {
-      await verifyMPTProof(utf8ToBytes('key2bb'), proof)
+      await verifyMerkleProof(utf8ToBytes('key2bb'), proof)
       assert.fail('expected error: Invalid proof provided')
     } catch (e: any) {
       assert.equal(e.message, 'Invalid proof provided')
@@ -69,7 +69,7 @@ describe('simple merkle proofs generation and verification', () => {
     // a valid exclusion proof then making it non-null
     myKey = utf8ToBytes('anyRandomKey')
     proof = await createMerkleProof(trie, myKey)
-    val = await verifyMPTProof(myKey, proof)
+    val = await verifyMerkleProof(myKey, proof)
     assert.equal(val, null, 'Expected value to be null')
     // now make the key non-null so the exclusion proof becomes invalid
     await trie.put(myKey, utf8ToBytes('thisIsaValue'))
@@ -87,7 +87,7 @@ describe('simple merkle proofs generation and verification', () => {
     await trie.put(utf8ToBytes('key1aa'), utf8ToBytes('0123456789012345678901234567890123456789xx'))
 
     const proof = await createMerkleProof(trie, utf8ToBytes('key1aa'))
-    const val = await verifyMPTProof(utf8ToBytes('key1aa'), proof)
+    const val = await verifyMerkleProof(utf8ToBytes('key1aa'), proof)
     assert.equal(bytesToUtf8(val!), '0123456789012345678901234567890123456789xx')
   })
 
@@ -97,7 +97,7 @@ describe('simple merkle proofs generation and verification', () => {
     await trie.put(utf8ToBytes('key1aa'), utf8ToBytes('01234'))
 
     const proof = await createMerkleProof(trie, utf8ToBytes('key1aa'))
-    const val = await verifyMPTProof(utf8ToBytes('key1aa'), proof)
+    const val = await verifyMerkleProof(utf8ToBytes('key1aa'), proof)
     assert.equal(bytesToUtf8(val!), '01234')
   })
 
@@ -118,15 +118,15 @@ describe('simple merkle proofs generation and verification', () => {
     await trie.put(utf8ToBytes('key3'), utf8ToBytes('1234567890123456789012345678901'))
 
     let proof = await createMerkleProof(trie, utf8ToBytes('key1'))
-    let val = await verifyMPTProof(utf8ToBytes('key1'), proof)
+    let val = await verifyMerkleProof(utf8ToBytes('key1'), proof)
     assert.equal(bytesToUtf8(val!), '0123456789012345678901234567890123456789Very_Long')
 
     proof = await createMerkleProof(trie, utf8ToBytes('key2'))
-    val = await verifyMPTProof(utf8ToBytes('key2'), proof)
+    val = await verifyMerkleProof(utf8ToBytes('key2'), proof)
     assert.equal(bytesToUtf8(val!), 'short')
 
     proof = await createMerkleProof(trie, utf8ToBytes('key3'))
-    val = await verifyMPTProof(utf8ToBytes('key3'), proof)
+    val = await verifyMerkleProof(utf8ToBytes('key3'), proof)
     assert.equal(bytesToUtf8(val!), '1234567890123456789012345678901')
   })
 
@@ -138,15 +138,15 @@ describe('simple merkle proofs generation and verification', () => {
     await trie.put(utf8ToBytes('c'), utf8ToBytes('c'))
 
     let proof = await createMerkleProof(trie, utf8ToBytes('a'))
-    let val = await verifyMPTProof(utf8ToBytes('a'), proof)
+    let val = await verifyMerkleProof(utf8ToBytes('a'), proof)
     assert.equal(bytesToUtf8(val!), 'a')
 
     proof = await createMerkleProof(trie, utf8ToBytes('b'))
-    val = await verifyMPTProof(utf8ToBytes('b'), proof)
+    val = await verifyMerkleProof(utf8ToBytes('b'), proof)
     assert.equal(bytesToUtf8(val!), 'b')
 
     proof = await createMerkleProof(trie, utf8ToBytes('c'))
-    val = await verifyMPTProof(utf8ToBytes('c'), proof)
+    val = await verifyMerkleProof(utf8ToBytes('c'), proof)
     assert.equal(bytesToUtf8(val!), 'c')
   })
