+++ packages/statemanager/src/proofs/merkle.ts
@@ -2,7 +2,7 @@ import {
   createMPTFromProof,
   createMerkleProof,
   updateMPTFromMerkleProof,
-  verifyMPTProof,
+  verifyMerkleProof,
 } from '@ethereumjs/mpt'
 import { RLP } from '@ethereumjs/rlp'
 import {
@@ -194,7 +194,7 @@ export async function verifyMerkleStateProof(
 
   // This returns the account if the proof is valid.
   // Verify that it matches the reported account.
-  const value = await verifyMPTProof(key, accountProof, {
+  const value = await verifyMerkleProof(key, accountProof, {
     useKeyHashing: true,
   })
 
@@ -240,7 +240,7 @@ export async function verifyMerkleStateProof(
     const storageProof = stProof.proof.map((value: PrefixedHexString) => hexToBytes(value))
     const storageValue = setLengthLeft(hexToBytes(stProof.value), 32)
     const storageKey = hexToBytes(stProof.key)
-    const proofValue = await verifyMPTProof(storageKey, storageProof, {
+    const proofValue = await verifyMerkleProof(storageKey, storageProof, {
       useKeyHashing: true,
     })
     const reportedValue = setLengthLeft(
